refactor: add useAccountContext hook as part of AccountProvider
