+++ src/features/stake/useStaking.ts
@@ -1,12 +1,12 @@
 import { useCallback } from 'react';
 import { fromCeloWei, toCeloWei } from 'src/formatters/amount';
-import { useAccount } from 'src/hooks/useAccount';
 import { useContracts } from 'src/hooks/useContracts';
+import { useAccountContext } from 'src/providers/AccountProvider';
 import { useExchangeContext } from 'src/providers/ExchangeProvider';
 import { Celo, CeloWei } from 'src/types/units';
 
 export function useStaking() {
-  const { address, loadBalances } = useAccount();
+  const { address, loadBalances } = useAccountContext();
   const { managerContract } = useContracts();
   const { celoExchangeRate } = useExchangeContext();
