+++ src/features/unstake/useUnstaking.ts
@@ -1,14 +1,14 @@
 import { useCallback, useEffect, useState } from 'react';
 import { GAS_LIMIT, GAS_PRICE } from 'src/config/consts';
 import { fromStCeloWei, toStCeloWei } from 'src/formatters/amount';
-import { useAccount } from 'src/hooks/useAccount';
 import { useContracts } from 'src/hooks/useContracts';
+import { useAccountContext } from 'src/providers/AccountProvider';
 import { useExchangeContext } from 'src/providers/ExchangeProvider';
 import { Celo, StCelo, StCeloWei } from 'src/types/units';
 import { PendingWithdrawal } from './types';
 
 export function useUnstaking() {
-  const { address, loadBalances } = useAccount();
+  const { address, loadBalances } = useAccountContext();
   const { managerContract, accountContract } = useContracts();
   const { stCeloExchangeRate } = useExchangeContext();
