--- src/hooks/useAccount.ts
@@ -1,15 +1,0 @@
-import { useContext } from 'react';
-import { AccountContext } from 'src/providers/AccountProvider';
-
-export function useAccount() {
-  const { isConnected, address, celoBalance, stCeloBalance, loadBalances } =
-    useContext(AccountContext);
-
-  return {
-    isConnected,
-    address,
-    celoBalance,
-    stCeloBalance,
-    loadBalances,
-  };
-}
