+++ src/providers/AccountProvider.tsx
@@ -1,6 +1,6 @@
-import { useCelo } from '@celo/react-celo';
-import { createContext, PropsWithChildren, useCallback, useEffect, useState } from 'react';
-import { useContracts } from 'src/hooks/useContracts';
+import { createContext, PropsWithChildren, useContext } from 'react';
+import { useAddress } from 'src/hooks/useAddress';
+import { useBalances } from 'src/hooks/useBalances';
 import { CeloWei, StCeloWei } from 'src/types/units';
 
 interface AccountContext {
@@ -10,54 +10,6 @@ interface AccountContext {
   stCeloBalance: StCeloWei;
   loadBalances: () => Promise<void>;
 }
-
-const useAddress = () => {
-  const { address: _address } = useCelo();
-  const address = _address ?? undefined;
-  const isConnected = !!address;
-  return { isConnected, address };
-};
-
-const useBalances = () => {
-  const { kit } = useCelo();
-  const { address } = useAddress();
-  const { stCeloContract } = useContracts();
-
-  const [celoBalance, setCeloBalance] = useState(new CeloWei(0));
-  const [stCeloBalance, setStCeloBalance] = useState(new StCeloWei(0));
-
-  const loadCeloBalance = useCallback(async () => {
-    const { eth } = kit.connection.web3;
-    if (!address) return;
-
-    const balance = await eth.getBalance(address);
-
-    setCeloBalance(new CeloWei(balance));
-  }, [kit.connection, address]);
-
-  const loadStCeloBalance = useCallback(async () => {
-    const stCeloBalance = await stCeloContract.methods.balanceOf(address).call({
-      from: address,
-    });
-    setStCeloBalance(new StCeloWei(stCeloBalance));
-  }, [address, stCeloContract]);
-
-  const loadBalances = useCallback(async () => {
-    await Promise.all([loadCeloBalance(), loadStCeloBalance()]);
-  }, [loadCeloBalance, loadStCeloBalance]);
-
-  useEffect(() => {
-    if (!address) return;
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    loadBalances();
-  }, [address, loadBalances]);
-
-  return {
-    celoBalance,
-    stCeloBalance,
-    loadBalances,
-  };
-};
 
 export const AccountContext = createContext<AccountContext>({
   isConnected: false,
@@ -85,3 +37,16 @@ export const AccountProvider = ({ children }: PropsWithChildren) => {
     </AccountContext.Provider>
   );
 };
+
+export function useAccountContext() {
+  const { isConnected, address, celoBalance, stCeloBalance, loadBalances } =
+    useContext(AccountContext);
+
+  return {
+    isConnected,
+    address,
+    celoBalance,
+    stCeloBalance,
+    loadBalances,
+  };
+}
